pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tuner-pipecat-sdk"
version = "0.1.1"
description = "Pipecat observer SDK for capturing latency, transcript, and usage data from pipecat and pipecat-flows pipelines."
readme = "README.md"
authors = [
    { name = "Tuner Team" },
]
# Pipecat (onnxruntime, numba, etc.) does not resolve on 3.14 yet; see README.
requires-python = ">=3.10,<3.14"
keywords = ["pipecat", "voice", "sdk", "observability", "agent"]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "pipecat-ai>=0.0.105",
    "pydantic>=2.0",
    "loguru~=0.7.2",
    "httpx>=0.27",
]

[project.urls]
Homepage = "https://github.com/usetuner/tuner-pipecat-sdk-python"
Documentation = "https://github.com/usetuner/tuner-pipecat-sdk-python/tree/main/docs"
Repository = "https://github.com/usetuner/tuner-pipecat-sdk-python"
Issues = "https://github.com/usetuner/tuner-pipecat-sdk-python/issues"

[project.optional-dependencies]
flows = [
    "pipecat-ai-flows>=0.0.24",
]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
]
lint = [
    "ruff>=0.6",
]
type = [
    "mypy>=1.10",
]
publish = [
    "hatch>=1.12",
    "twine>=5.0",
]
dev = [
    "tuner-pipecat-sdk[test,lint,type,publish,flows]",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-ra --strict-markers --strict-config --cov=tuner_pipecat_sdk --cov-report=term-missing"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/test_observer.py" = ["E402"]

[tool.mypy]
python_version = "3.10"
packages = ["tuner_pipecat_sdk"]
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
strict_equality = true
no_implicit_optional = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["pipecat.*","pipecat_flows.*"]
ignore_missing_imports = true

[tool.coverage.run]
source = ["tuner_pipecat_sdk"]
branch = true
omit = [
    "src/tuner_pipecat_sdk/observer.py",
    "src/tuner_pipecat_sdk/flow_manager_integration.py",
]

[tool.coverage.report]
show_missing = true
skip_covered = false
fail_under = 85

[tool.hatch.build.targets.wheel]
packages = ["src/tuner_pipecat_sdk"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
